+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added `config.ExcludeGRPCCodesForPessimization()` opttion for exclude some grpc codes from pessimization rules 
 * Refactored pessimization node conditions
 * Added closing of ticker in `conn.Conn.connParker`
 * Removed `config.WithSharedPool` and usages it
