+++ config/config.go
@@ -8,10 +8,12 @@ import (
 	"time"
 
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/balancers"
 	"github.com/ydb-platform/ydb-go-sdk/v3/credentials"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/balancer"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/meta"
 	builder "github.com/ydb-platform/ydb-go-sdk/v3/internal/net"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/resolver"
@@ -76,26 +78,34 @@ type Config interface {
 
 	// UseDNSResolver is a flag about using dns-resolving or not
 	UseDNSResolver() bool
+
+	// ExcludeGRPCCodesForPessimization defines grpc codes for exclude its from pessimization trigger
+	ExcludeGRPCCodesForPessimization() []errors.TransportErrorCode
 }
 
 // Config contains driver configuration options.
 type config struct {
-	trace                trace.Driver
-	operationTimeout     time.Duration
-	operationCancelAfter time.Duration
-	dialTimeout          time.Duration
-	connectionTTL        time.Duration
-	balancer             balancer.Balancer
-	secure               bool
-	dnsResolver          bool
-	endpoint             string
-	database             string
-	requestsType         string
-	userAgent            string
-	grpcOptions          []grpc.DialOption
-	credentials          credentials.Credentials
-	tlsConfig            *tls.Config
-	meta                 meta.Meta
+	trace                            trace.Driver
+	operationTimeout                 time.Duration
+	operationCancelAfter             time.Duration
+	dialTimeout                      time.Duration
+	connectionTTL                    time.Duration
+	balancer                         balancer.Balancer
+	secure                           bool
+	dnsResolver                      bool
+	endpoint                         string
+	database                         string
+	requestsType                     string
+	userAgent                        string
+	excludeGRPCCodesForPessimization []errors.TransportErrorCode
+	grpcOptions                      []grpc.DialOption
+	credentials                      credentials.Credentials
+	tlsConfig                        *tls.Config
+	meta                             meta.Meta
+}
+
+func (c *config) ExcludeGRPCCodesForPessimization() []errors.TransportErrorCode {
+	return c.excludeGRPCCodesForPessimization
 }
 
 func (c *config) UseDNSResolver() bool {
@@ -262,6 +272,20 @@ func WithTLSSInsecureSkipVerify() Option {
 func WithGrpcOptions(option ...grpc.DialOption) Option {
 	return func(c *config) {
 		c.grpcOptions = append(c.grpcOptions, option...)
+	}
+}
+
+func ExcludeGRPCCodesForPessimization(codes ...codes.Code) Option {
+	return func(c *config) {
+		c.excludeGRPCCodesForPessimization = append(
+			c.excludeGRPCCodesForPessimization,
+			func() (teCodes []errors.TransportErrorCode) {
+				for _, c := range codes {
+					teCodes = append(teCodes, errors.FromGRPCCode(c))
+				}
+				return teCodes
+			}()...,
+		)
 	}
 }
