+++ internal/errors/issues.go
@@ -1,6 +1,10 @@
 package errors
 
-import "bytes"
+import (
+	"bytes"
+
+	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Issue"
+)
 
 // NewWithIssues returns error which contains child issues
 func NewWithIssues(text string, issues ...error) error {
@@ -29,4 +33,29 @@ func (e *errorWithIssues) Error() string {
 	}
 	b.WriteString("]")
 	return b.String()
+}
+
+// Issue struct
+type Issue struct {
+	Message  string
+	Code     uint32
+	Severity uint32
+}
+
+type IssueIterator []*Ydb_Issue.IssueMessage
+
+func (it IssueIterator) Len() int {
+	return len(it)
+}
+
+func (it IssueIterator) Get(i int) (issue Issue, nested IssueIterator) {
+	x := it[i]
+	if xs := x.Issues; len(xs) > 0 {
+		nested = IssueIterator(xs)
+	}
+	return Issue{
+		Message:  x.GetMessage(),
+		Code:     x.GetIssueCode(),
+		Severity: x.GetSeverity(),
+	}, nested
 }
