+++ internal/errors/operation.go
@@ -8,6 +8,79 @@ import (
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Issue"
 )
 
+// StatusCode reports unsuccessful operation status code.
+type StatusCode int32
+
+func (e StatusCode) String() string {
+	return Ydb.StatusIds_StatusCode_name[int32(e)]
+}
+
+// Errors describing unsusccessful operation status.
+const (
+	StatusUnknownStatus      = StatusCode(Ydb.StatusIds_STATUS_CODE_UNSPECIFIED)
+	StatusBadRequest         = StatusCode(Ydb.StatusIds_BAD_REQUEST)
+	StatusUnauthorized       = StatusCode(Ydb.StatusIds_UNAUTHORIZED)
+	StatusInternalError      = StatusCode(Ydb.StatusIds_INTERNAL_ERROR)
+	StatusAborted            = StatusCode(Ydb.StatusIds_ABORTED)
+	StatusUnavailable        = StatusCode(Ydb.StatusIds_UNAVAILABLE)
+	StatusOverloaded         = StatusCode(Ydb.StatusIds_OVERLOADED)
+	StatusSchemeError        = StatusCode(Ydb.StatusIds_SCHEME_ERROR)
+	StatusGenericError       = StatusCode(Ydb.StatusIds_GENERIC_ERROR)
+	StatusTimeout            = StatusCode(Ydb.StatusIds_TIMEOUT)
+	StatusBadSession         = StatusCode(Ydb.StatusIds_BAD_SESSION)
+	StatusPreconditionFailed = StatusCode(Ydb.StatusIds_PRECONDITION_FAILED)
+	StatusAlreadyExists      = StatusCode(Ydb.StatusIds_ALREADY_EXISTS)
+	StatusNotFound           = StatusCode(Ydb.StatusIds_NOT_FOUND)
+	StatusSessionExpired     = StatusCode(Ydb.StatusIds_SESSION_EXPIRED)
+	StatusCancelled          = StatusCode(Ydb.StatusIds_CANCELLED)
+	StatusUndetermined       = StatusCode(Ydb.StatusIds_UNDETERMINED)
+	StatusUnsupported        = StatusCode(Ydb.StatusIds_UNSUPPORTED)
+	StatusSessionBusy        = StatusCode(Ydb.StatusIds_SESSION_BUSY)
+)
+
+func statusCode(s Ydb.StatusIds_StatusCode) StatusCode {
+	switch s {
+	case Ydb.StatusIds_BAD_REQUEST:
+		return StatusBadRequest
+	case Ydb.StatusIds_UNAUTHORIZED:
+		return StatusUnauthorized
+	case Ydb.StatusIds_INTERNAL_ERROR:
+		return StatusInternalError
+	case Ydb.StatusIds_ABORTED:
+		return StatusAborted
+	case Ydb.StatusIds_UNAVAILABLE:
+		return StatusUnavailable
+	case Ydb.StatusIds_OVERLOADED:
+		return StatusOverloaded
+	case Ydb.StatusIds_SCHEME_ERROR:
+		return StatusSchemeError
+	case Ydb.StatusIds_GENERIC_ERROR:
+		return StatusGenericError
+	case Ydb.StatusIds_TIMEOUT:
+		return StatusTimeout
+	case Ydb.StatusIds_BAD_SESSION:
+		return StatusBadSession
+	case Ydb.StatusIds_PRECONDITION_FAILED:
+		return StatusPreconditionFailed
+	case Ydb.StatusIds_ALREADY_EXISTS:
+		return StatusAlreadyExists
+	case Ydb.StatusIds_NOT_FOUND:
+		return StatusNotFound
+	case Ydb.StatusIds_SESSION_EXPIRED:
+		return StatusSessionExpired
+	case Ydb.StatusIds_CANCELLED:
+		return StatusCancelled
+	case Ydb.StatusIds_UNDETERMINED:
+		return StatusUndetermined
+	case Ydb.StatusIds_UNSUPPORTED:
+		return StatusUnsupported
+	case Ydb.StatusIds_SESSION_BUSY:
+		return StatusSessionBusy
+	default:
+		return StatusUnknownStatus
+	}
+}
+
 // OperationError reports about operation fail.
 type OperationError struct {
 	Reason StatusCode
