+++ internal/errors/transport.go
@@ -10,34 +10,8 @@ import (
 	grpcCodes "google.golang.org/grpc/codes"
 	grpcStatus "google.golang.org/grpc/status"
 
-	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Issue"
 )
-
-// Issue struct
-type Issue struct {
-	Message  string
-	Code     uint32
-	Severity uint32
-}
-
-type IssueIterator []*Ydb_Issue.IssueMessage
-
-func (it IssueIterator) Len() int {
-	return len(it)
-}
-
-func (it IssueIterator) Get(i int) (issue Issue, nested IssueIterator) {
-	x := it[i]
-	if xs := x.Issues; len(xs) > 0 {
-		nested = IssueIterator(xs)
-	}
-	return Issue{
-		Message:  x.GetMessage(),
-		Code:     x.GetIssueCode(),
-		Severity: x.GetSeverity(),
-	}, nested
-}
 
 type TransportError struct {
 	Reason TransportErrorCode
@@ -152,79 +126,6 @@ func dumpIssues(buf *bytes.Buffer, ms []*Ydb_Issue.IssueMessage) {
 	}
 }
 
-// StatusCode reports unsuccessful operation status code.
-type StatusCode int32
-
-func (e StatusCode) String() string {
-	return Ydb.StatusIds_StatusCode_name[int32(e)]
-}
-
-// Errors describing unsusccessful operation status.
-const (
-	StatusUnknownStatus      = StatusCode(Ydb.StatusIds_STATUS_CODE_UNSPECIFIED)
-	StatusBadRequest         = StatusCode(Ydb.StatusIds_BAD_REQUEST)
-	StatusUnauthorized       = StatusCode(Ydb.StatusIds_UNAUTHORIZED)
-	StatusInternalError      = StatusCode(Ydb.StatusIds_INTERNAL_ERROR)
-	StatusAborted            = StatusCode(Ydb.StatusIds_ABORTED)
-	StatusUnavailable        = StatusCode(Ydb.StatusIds_UNAVAILABLE)
-	StatusOverloaded         = StatusCode(Ydb.StatusIds_OVERLOADED)
-	StatusSchemeError        = StatusCode(Ydb.StatusIds_SCHEME_ERROR)
-	StatusGenericError       = StatusCode(Ydb.StatusIds_GENERIC_ERROR)
-	StatusTimeout            = StatusCode(Ydb.StatusIds_TIMEOUT)
-	StatusBadSession         = StatusCode(Ydb.StatusIds_BAD_SESSION)
-	StatusPreconditionFailed = StatusCode(Ydb.StatusIds_PRECONDITION_FAILED)
-	StatusAlreadyExists      = StatusCode(Ydb.StatusIds_ALREADY_EXISTS)
-	StatusNotFound           = StatusCode(Ydb.StatusIds_NOT_FOUND)
-	StatusSessionExpired     = StatusCode(Ydb.StatusIds_SESSION_EXPIRED)
-	StatusCancelled          = StatusCode(Ydb.StatusIds_CANCELLED)
-	StatusUndetermined       = StatusCode(Ydb.StatusIds_UNDETERMINED)
-	StatusUnsupported        = StatusCode(Ydb.StatusIds_UNSUPPORTED)
-	StatusSessionBusy        = StatusCode(Ydb.StatusIds_SESSION_BUSY)
-)
-
-func statusCode(s Ydb.StatusIds_StatusCode) StatusCode {
-	switch s {
-	case Ydb.StatusIds_BAD_REQUEST:
-		return StatusBadRequest
-	case Ydb.StatusIds_UNAUTHORIZED:
-		return StatusUnauthorized
-	case Ydb.StatusIds_INTERNAL_ERROR:
-		return StatusInternalError
-	case Ydb.StatusIds_ABORTED:
-		return StatusAborted
-	case Ydb.StatusIds_UNAVAILABLE:
-		return StatusUnavailable
-	case Ydb.StatusIds_OVERLOADED:
-		return StatusOverloaded
-	case Ydb.StatusIds_SCHEME_ERROR:
-		return StatusSchemeError
-	case Ydb.StatusIds_GENERIC_ERROR:
-		return StatusGenericError
-	case Ydb.StatusIds_TIMEOUT:
-		return StatusTimeout
-	case Ydb.StatusIds_BAD_SESSION:
-		return StatusBadSession
-	case Ydb.StatusIds_PRECONDITION_FAILED:
-		return StatusPreconditionFailed
-	case Ydb.StatusIds_ALREADY_EXISTS:
-		return StatusAlreadyExists
-	case Ydb.StatusIds_NOT_FOUND:
-		return StatusNotFound
-	case Ydb.StatusIds_SESSION_EXPIRED:
-		return StatusSessionExpired
-	case Ydb.StatusIds_CANCELLED:
-		return StatusCancelled
-	case Ydb.StatusIds_UNDETERMINED:
-		return StatusUndetermined
-	case Ydb.StatusIds_UNSUPPORTED:
-		return StatusUnsupported
-	case Ydb.StatusIds_SESSION_BUSY:
-		return StatusSessionBusy
-	default:
-		return StatusUnknownStatus
-	}
-}
-
 type TransportErrorCode int32
 
 func (t TransportErrorCode) String() string {
@@ -312,7 +213,7 @@ var grpcCodesToTransportError = [...]TransportErrorCode{
 	grpcCodes.Unauthenticated:    TransportErrorUnauthenticated,
 }
 
-func transportErrorCode(c grpcCodes.Code) TransportErrorCode {
+func FromGRPCCode(c grpcCodes.Code) TransportErrorCode {
 	if int(c) < len(grpcCodesToTransportError) {
 		return grpcCodesToTransportError[c]
 	}
@@ -378,7 +279,7 @@ func FromGRPCError(err error, opts ...teOpt) error {
 
 	if s, ok := grpcStatus.FromError(err); ok {
 		te := &TransportError{
-			Reason:  transportErrorCode(s.Code()),
+			Reason:  FromGRPCCode(s.Code()),
 			message: s.Message(),
 			err:     s.Err(),
 			details: s.Details(),
